feat: add BayesianProbabilityPublisher service and integrate with MultiScenarioLocator
